requirements.txt
asgiref==3.2.7
certifi==2019.11.28
chardet==3.0.4
Django==3.0.4
gunicorn==20.0.4
idna==2.9
pytz==2019.3
requests==2.23.0
sqlparse==0.3.1
tqdm==4.44.1
urllib3==1.25.8
django-environ==0.4.5
psycopg2==2.8.4
ijson==2.6.1
python-slugify==4.0.0
djangorestframework==3.11.0
django-extra-fields==2.0.3
django-filter==2.2.0